Add slow sync log rolling test in TestAsyncLogRolling.
